Simplify upload dialog

+++ PatternPaint/controllers/blinkyuploader.h
@@ -51,8 +51,8 @@ public:
     virtual QList<PatternWriter::Encoding> getSupportedEncodings() const = 0;
 
 signals:
-    /// Sends an update about the upload progress, from 0 to 1
-    void progressChanged(float progress);
+    /// Sends an update about the upload progress, from 0 to 100
+    void progressChanged(int progress);
 
     /// Sends a signal at end of upload to report the result.
     void finished(bool result);
